initial chatflow split wip
